third_party/blink/web_tests/platform/webview/external/wpt/push-api/idlharness.https.any.serviceworker-expected.txt

This is a testharness.js-based test.
Found 66 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] PushManager interface: existence and properties of interface object
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager interface object length
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager interface object name
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager interface: attribute supportedContentEncodings
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager interface: operation subscribe(optional PushSubscriptionOptionsInit)
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager interface: operation getSubscription()
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager interface: operation permissionState(optional PushSubscriptionOptionsInit)
  assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
[FAIL] PushManager must be primary interface of registration.pushManager
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of registration.pushManager
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PushManager interface: registration.pushManager must inherit property "supportedContentEncodings" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PushManager interface: registration.pushManager must inherit property "subscribe(optional PushSubscriptionOptionsInit)" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PushManager interface: calling subscribe(optional PushSubscriptionOptionsInit) on registration.pushManager with too few arguments must throw TypeError
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PushManager interface: registration.pushManager must inherit property "getSubscription()" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PushManager interface: registration.pushManager must inherit property "permissionState(optional PushSubscriptionOptionsInit)" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PushManager interface: calling permissionState(optional PushSubscriptionOptionsInit) on registration.pushManager with too few arguments must throw TypeError
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PushSubscriptionOptions interface: existence and properties of interface object
  assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
[FAIL] PushSubscriptionOptions interface object length
  assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
[FAIL] PushSubscriptionOptions interface object name
  assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
[FAIL] PushSubscriptionOptions interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
[FAIL] PushSubscriptionOptions interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
[FAIL] PushSubscriptionOptions interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
[FAIL] PushSubscriptionOptions interface: attribute userVisibleOnly
  assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
[FAIL] PushSubscriptionOptions interface: attribute applicationServerKey
  assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
[FAIL] PushSubscription interface: existence and properties of interface object
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface object length
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface object name
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface: attribute endpoint
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface: attribute expirationTime
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface: attribute options
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface: operation getKey(PushEncryptionKeyName)
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface: operation unsubscribe()
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushSubscription interface: operation toJSON()
  assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
[FAIL] PushEvent interface: existence and properties of interface object
  assert_own_property: self does not have own property "PushEvent" expected property "PushEvent" missing
[FAIL] PushEvent interface object length
  assert_own_property: self does not have own property "PushEvent" expected property "PushEvent" missing
[FAIL] PushEvent interface object name
  assert_own_property: self does not have own property "PushEvent" expected property "PushEvent" missing
[FAIL] PushEvent interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "PushEvent" expected property "PushEvent" missing
[FAIL] PushEvent interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "PushEvent" expected property "PushEvent" missing
[FAIL] PushEvent interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "PushEvent" expected property "PushEvent" missing
[FAIL] PushEvent interface: attribute data
  assert_own_property: self does not have own property "PushEvent" expected property "PushEvent" missing
[FAIL] PushEvent must be primary interface of new PushEvent("type")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PushEvent is not defined"
[FAIL] Stringification of new PushEvent("type")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PushEvent is not defined"
[FAIL] PushEvent interface: new PushEvent("type") must inherit property "data" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PushEvent is not defined"
[FAIL] PushSubscriptionChangeEvent interface: existence and properties of interface object
  assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
[FAIL] PushSubscriptionChangeEvent interface object length
  assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
[FAIL] PushSubscriptionChangeEvent interface object name
  assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
[FAIL] PushSubscriptionChangeEvent interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
[FAIL] PushSubscriptionChangeEvent interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
[FAIL] PushSubscriptionChangeEvent interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
[FAIL] PushSubscriptionChangeEvent interface: attribute newSubscription
  assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
[FAIL] PushSubscriptionChangeEvent interface: attribute oldSubscription
  assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
[FAIL] PushSubscriptionChangeEvent must be primary interface of new PushSubscriptionChangeEvent("pushsubscriptionchange")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PushSubscriptionChangeEvent is not defined"
[FAIL] Stringification of new PushSubscriptionChangeEvent("pushsubscriptionchange")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PushSubscriptionChangeEvent is not defined"
[FAIL] PushSubscriptionChangeEvent interface: new PushSubscriptionChangeEvent("pushsubscriptionchange") must inherit property "newSubscription" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PushSubscriptionChangeEvent is not defined"
[FAIL] PushSubscriptionChangeEvent interface: new PushSubscriptionChangeEvent("pushsubscriptionchange") must inherit property "oldSubscription" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: PushSubscriptionChangeEvent is not defined"
[FAIL] ServiceWorkerRegistration interface: attribute pushManager
  assert_true: The prototype object must have a property "pushManager" expected true got false
[FAIL] ServiceWorkerRegistration interface: registration must inherit property "pushManager" with the proper type
  assert_inherits: property "pushManager" not found in prototype chain
[FAIL] ServiceWorkerGlobalScope interface: attribute onpush
  assert_own_property: The global object must have a property "onpush" expected property "onpush" missing
[FAIL] ServiceWorkerGlobalScope interface: attribute onpushsubscriptionchange
  assert_own_property: The global object must have a property "onpushsubscriptionchange" expected property "onpushsubscriptionchange" missing
[FAIL] ServiceWorkerGlobalScope interface: self must inherit property "onpush" with the proper type
  assert_own_property: expected property "onpush" missing
[FAIL] ServiceWorkerGlobalScope interface: self must inherit property "onpushsubscriptionchange" with the proper type
  assert_own_property: expected property "onpushsubscriptionchange" missing
Harness: the test ran to completion.
